% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPedigreeFunctions.R
\name{affected_onlyPed}
\alias{affected_onlyPed}
\title{Reduce pedigree to contain only disease-affecteds, obligate carriers, and founders.}
\usage{
affected_onlyPed(ped_file)
}
\arguments{
\item{ped_file}{An object of class \code{ped}, inherits from a data.frame. The pedigree to reduce to disease-affected relatives.}
}
\value{
\code{retA_ped} A pedigree containing only affected members, obligate carriers, and founders.
}
\description{
Reduce pedigree to contain only disease-affecteds, obligate carriers, and founders.
}
\keyword{internal}
