% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Macro_SimFunctions.R
\name{allcombos_addCarProb}
\alias{allcombos_addCarProb}
\title{Inocrporate the carrier probability into the probability of configuraiton}
\usage{
allcombos_addCarProb(ped, like_mat, carrier_prob)
}
\arguments{
\item{ped}{An object of class \code{ped}. A pedigree generated by \code{sim_ped} or \code{sim_RVped}, or an object created by the function \code{new.ped}, for details please see documantation and vignette for package \code{SimRVPedigree}.}

\item{carrier_prob}{The cumulative carrier probabilty of all crvs as a group.}
}
\value{
A list containing the following:

\item{\code{like_mat} }{A matrix of values that represent the likelihood given a configuration (column) and relized value of tau (row).}

\item{\code{configs} }{A matrix of sharing configurations.  The columns are named for the IDs of the disease-affected relatives in the pedigree. When \code{TRUE} the individual is a carrier of the cRV, when \code{FALSE} the individual is not a carrier.}
}
\description{
Compute sharing probabilities for every global configuration (includes the zero configuration) and initialization of tau.  The probability of the zero configuration will be used to condition on the event that the RV was observed in at least one study member.  This function requires an assummed value for the carrier probability of the RV
}
\examples{
library(RVMethods)
data(study_pedigrees)

library(SimRVPedigree)
plot(study_pedigrees[study_pedigrees$FamID == 58, ])
myCombos = test_allcombos_includeZero(ped = study_pedigrees[study_pedigrees$FamID == 58, ],
                                      subtypes = c("HL", "NHL"),
                                      tau_grid = make_tauGrid(),
                                      carrier_prob = 0.002)

myCombos

rowSums(myCombos$like_mat)

}
\keyword{internal}
