% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPedigreeFunctions.R
\name{assign_obgcarrier_tau}
\alias{assign_obgcarrier_tau}
\title{Assign tau to obligate carriers}
\usage{
assign_obgcarrier_tau(id, parents, typeA, typeB)
}
\arguments{
\item{id}{the id (index) of the obligate carrier}

\item{parents}{the parent matrix}

\item{typeA}{the list of indicies of the type A's}

\item{typeB}{the list of indicies of the type B's}
}
\value{
character, either "A" or "B
}
\description{
Assign tau to obligate carriers
}
\keyword{internal}
