% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyAnalysisFunctions.R
\name{compute_binID}
\alias{compute_binID}
\title{Compute the binID for the observed familial genotypes}
\usage{
compute_binID(affected_genotypes_at_marker)
}
\arguments{
\item{affected_genotypes_at_marker}{The genotype for the disease affected relatives at a marker.}
}
\value{
The binID, which is the base 2 representation of the sharing configuration.
}
\description{
Compute the binID for the observed familial genotypes
}
\details{
Note that under our model, cRVs may only be introduced by a single founder. If an affected relative has 2 copies of the variant then it must have been introduced by multiple founders.  When this occurs, we set the  the binID to 0 since these variants are not applicable under our model.
}
\keyword{internal}
