% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRDistribution_Functions.R
\name{compute_distributions}
\alias{compute_distributions}
\title{Compute prioritization statistics and probabilities}
\usage{
compute_distributions(peds, subtypes, carrier_prob = 0.002,
  tau_increment = 0.05, subtype_weights = NULL, useK = FALSE)
}
\arguments{
\item{peds}{A ped object containing two or more pedigrees.}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}

\item{carrier_prob}{The cumulative carrier probabilty of all cRVs as a group.}

\item{tau_increment}{Numeric. The width of the grid for the taus}

\item{subtype_weights}{A vector of length 2.  When assuming an informative prior for the founders; the weights the individuals with subtypes A and B, respectively.  By default, \code{subtype_weights = NULL} so that no founder weights are applied (i.e. flat founder prior).}
}
\value{
A list of three \code{data.frames}: \code{GlobalDist}, \code{GlobalTransDist}, and \code{LocalDist}; which include:

\item{RV status indicators}{Binary indicators of RV status for each disease-affected relative, notated as \code{FamID:ID}, where \code{FamID} is the family identification number and \code{ID} is the individual identification number.}

\item{L_max/L_null}{The likelihood under the alternative and null hypotheses}

\item{LR}{The likelihood ratio statistic}

\item{null_configProb}{The null probability of the sharing configuration.}

\item{tau_A/tau_B}{The values of tau_A and tau_B under the alternative hypothesis.}

\item{K}{The number of relatives that were observed to carry the RV}

\item{LR_pvalue}{the p-value if the LR statistic}

\item{binID}{a unique ID for each sharing configuration.}
}
\description{
Computes all considered prioritization statistics and probabilities
}
\examples{
library(RVMethods)
data(study_pedigrees)

library(SimRVPedigree)
plot(study_pedigrees[study_pedigrees$FamID == 304, ])
plot(study_pedigrees[study_pedigrees$FamID == 58, ])


FBSDists = compute_distributions(ped = study_pedigrees[study_pedigrees$FamID \%in\% c(58, 304), ],
                                 subtypes = c("HL", "NHL"),
                                 carrier_prob = 0.002)


}
