% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroConfig_Functions.R
\name{compute_famLR_zeroConfig}
\alias{compute_famLR_zeroConfig}
\title{Compute the sharing probablities under the null, i.e. when all taus are 1/2}
\usage{
compute_famLR_zeroConfig(famLike_results, tau_grid, ped, subtypes)
}
\arguments{
\item{famLike_results}{The output returned by the \code{\link{test_allcombos}} function.}

\item{ped}{An object of class \code{ped}. A pedigree generated by \code{sim_ped} or \code{sim_RVped}, or an object created by the function \code{new.ped}, for details please see documantation and vignette for package \code{SimRVPedigree}.}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}
}
\value{
A data frame containing the weights for each configuration in the pedigree. See Details.
}
\description{
Compute the sharing probablities under the null, i.e. when all taus are 1/2
}
\keyword{internal}
