% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalConfigFunctions.R
\name{compute_global_weight}
\alias{compute_global_weight}
\title{Compute Global weights for a set of families}
\usage{
compute_global_weight(likeGrids_byFam, weights_byFam, famID_index,
  tau_grid)
}
\arguments{
\item{likeGrids_byFam}{A list of results from test_allcombos, one for each family.}

\item{weights_byFam}{A list of results from compute_famWeights, one for each family.}

\item{famID_index}{The family IDs, indexed in the same order as likeGrids_byFam and weights_byFam}

\item{tau_grid}{The grid of tau values}
}
\value{
A data frame of global configurations anf weights for the set of provided families.
}
\description{
Compute Global weights for a set of families
}
\keyword{internal}
