% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight_functions.R
\name{compute_RVS_weight}
\alias{compute_RVS_weight}
\title{Compute the RVS weight}
\usage{
compute_RVS_weight(all_stats, config_index)
}
\arguments{
\item{all_stats}{The data frame of stats for all configurations}

\item{config_index}{The index of the observed confiugration}
}
\value{
the RVS weight
}
\description{
Compute the RVS weight
}
\keyword{internal}
