% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cn_Methods.R
\name{compute_sharingProb}
\alias{compute_sharingProb}
\title{Calculate the probability of an observed sharing configuration given specified transmission probabilities.}
\usage{
compute_sharingProb(ped, subtypes, tau, carriers = NULL,
  subtype_weights = NULL)
}
\arguments{
\item{ped}{An object of class \code{ped}.}

\item{subtypes}{A list of two subtypes that identify the subtype ranking.  The label for the more the genetically-complelling subtype is listed first.}

\item{tau}{A list of two transmission probabilites, with the transmission probability for the genetically-complelling disease-subtype listed first. These represent the transmission probability from a heterozygous parent to an offspring.}

\item{carriers}{subjects in pedigree share the variant}
}
\value{
The transmssion-based probability of the sharing configuration for the transmission probabilities specified in the argument \code{tau}.
}
\description{
Calculate the probability of an observed sharing configuration given specified transmission probabilities.
}
\keyword{internal}
