% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyAnalysisFunctions.R
\name{compute_studyWeights}
\alias{compute_studyWeights}
\title{Compute weights for variants shared among disease-affected study participants}
\usage{
compute_studyWeights(famStudy_obj, subtypes, tau_increment = 0.05,
  subtype_weights = NULL)
}
\arguments{
\item{famStudy_obj}{An object of class \code{famStudy}, i.e. returned by the \code{sim_RVstudy} function (included in the \code{SimRVSequences} package).}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}

\item{tau_increment}{Numeric. The width of the grid for the taus}

\item{subtype_weights}{A vector of length 2.  The weights to applied to individuals with the subtypes listed in \code{subtypes}. By default, \code{subtype_weights = NULL} so that no weights are applied.}
}
\value{
A dataframe cataloging all of the SNVs, with weights included as columns w_LR, w_RVS, and w_RVS2.  See \code{\link{compute_familyWeights}} documentation for weight definitions.
}
\description{
Compute weights for variants shared among disease-affected study participants
}
\details{
The parameter space of  \eqn{\tau_a} and \eqn{\tau_b} is given by
\eqn{\Omega_{con} = ((\tau_{a}, \tau_{b}): 0.5 \le \tau_{b} < \tau_{a} \le 1 )}.
}
\examples{
library(RVMethods)

# load example pedigrees
data(study_pedigrees)
head(study_pedigrees)

#load example population haplotype data
data(pop_fdat)
str(pop_fdat)

#simulate sequence data for pedigrees with the SimRVSequences package
library(SimRVSequences)
ped_seq = sim_RVstudy(ped_files = study_pedigrees,
                      SNV_data = pop_fdat)
class(ped_seq)
# for more information about the sim_RVstudy function and it's output,
# i.e. objects of class famStudy, please refer to the vigentte
# for the SimRVSequences R package


# compute weights for SNVs shared among the
# disease-affected relatives in the study
SNV_weights = compute_studyWeights(famStudy_obj = ped_seq,
                                   subtypes = c("HL", "NHL"))

head(SNV_weights)

}
