% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight_functions.R
\name{compute_transmission_weight}
\alias{compute_transmission_weight}
\title{Compute the transmission based weight}
\usage{
compute_transmission_weight(all_stats, config_index)
}
\arguments{
\item{all_stats}{The data frame of stats for all configurations}

\item{config_index}{The index of the observed confiugration}
}
\value{
the transmission based weight
}
\description{
Compute the transmission based weight
}
\keyword{internal}
