% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRDistribution_Functions.R
\name{conditioned_semiGlobalDist}
\alias{conditioned_semiGlobalDist}
\title{(Option 3) Determine the MLEs and LR in the semi global framework, with conditioning imposed.}
\usage{
conditioned_semiGlobalDist(likeGrids_byFam, famID_index, tau_grid)
}
\arguments{
\item{likeGrids_byFam}{A list of results from test_allcombos, one for each family.}

\item{famID_index}{The family IDs, indexed in the same order as likeGrids_byFam and weights_byFam}

\item{tau_grid}{The grid of tau values}

\item{weights_byFam}{A list of results from compute_famWeights, one for each family.}
}
\value{
A data frame of global configurations anf weights for the set of provided families.
}
\description{
This will be an internal function
}
\keyword{internal}
