% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cn_Methods.R
\name{denomProb}
\alias{denomProb}
\title{denominator of sharing probability}
\usage{
denomProb(net, procPed)
}
\arguments{
\item{procPed}{pedigree object that has been process with processPedigree}

\item{gRain}{bayesian network}
}
\value{
denominator value
}
\description{
calculates the denominator of the sharing probability
outline in section 2.1 of Bureau et al.
}
\keyword{internal}
