% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyAnalysisFunctions.R
\name{find_famSharing}
\alias{find_famSharing}
\title{Determine the families that share variants in the simulated data}
\usage{
find_famSharing(binID_mat)
}
\arguments{
\item{binID_mat}{The matrix of binIDs for the families in the study. The $(i,j)^{th}$ entry represents the binary representation of the configuration for the $j^{th}$ variant of the $i^{th}$ family.}
}
\value{
A list of sharing combinations, i.e. a list of FamID lists.
}
\description{
Determine the families that share variants in the simulated data
}
\keyword{internal}
