% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cn_Methods.R
\name{findDescendents}
\alias{findDescendents}
\title{Find all descendents of a pedigree member}
\usage{
findDescendents(id, parents)
}
\arguments{
\item{id}{Numeric. The ID of the member whose offspring we would like to identify}

\item{parents}{The parent matrix returned by ProcessPed}
}
\value{
a list of descendent IDs
}
\description{
Find all descendents of a pedigree member
}
\keyword{internal}
