% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cn_Methods.R
\name{findOffspring}
\alias{findOffspring}
\title{Find offspring of a pedigree member}
\usage{
findOffspring(id, parents)
}
\arguments{
\item{id}{Numeric. The ID of the member whose offspring we would like to identify}

\item{parents}{The parent matrix returned by ProcessPed}
}
\value{
a list of offspring IDs
}
\description{
Find offspring of a pedigree member
}
\keyword{internal}
