% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroConfig_Functions.R
\name{get_MLE_zeroConfig}
\alias{get_MLE_zeroConfig}
\title{Find the values of tau that maximize the likelihood function}
\usage{
get_MLE_zeroConfig(like_mat, tau_grid)
}
\arguments{
\item{like_mat}{The matrix of likelihood values given a sharing configuation and values of tau, i.e. the first item returned by \code{\link{test_allcombos}}.}

\item{tau_grid}{A grid of tau values produced by \code{expand.grid}.  \code{test_allcombos} will compute each sharing probability at every specified value of tau in \code{tau_grid}.}
}
\value{
A list of taus that maximize the likelihood for each configuration
}
\description{
Find the values of tau that maximize the likelihood function. NOTE: if there is not a unique MCLE, this funciton orders the results conservatively.  That is, the smallest values of tau_A and tau_B are the first values returned.
}
\keyword{internal}
