% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{make_tauGrid}
\alias{make_tauGrid}
\title{Create grid of tau values}
\usage{
make_tauGrid(increment_width = 0.05, constrained = TRUE)
}
\arguments{
\item{increment_width}{The distance between consecutive tau values}
}
\value{
tau_grid a matrix of taus
}
\description{
Create grid of tau values
}
\keyword{internal}
