% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cn_Methods.R
\name{numerProb}
\alias{numerProb}
\title{numerator of sharing probability}
\usage{
numerProb(net, procPed)
}
\arguments{
\item{procPed}{pedigree object that has been process with processPedigree}

\item{gRain}{bayesian network}
}
\value{
numerator value
}
\description{
calculates the numerator of the sharing probability
outline in section 2.1 of Bureau et al.
}
\keyword{internal}
