% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_fdat.R
\docType{data}
\name{pop_fdat}
\alias{pop_fdat}
\title{Example founder haplotype data}
\format{An object of class \code{SNVdata} (inherits from \code{list}) of length 3.}
\usage{
data(pop_fdat)
}
\description{
An object of class \code{SNVdata}, returned by the \code{read_slim} function included with the \code{SimRVSequences} R package.
}
\references{
Christina Nieuwoudt and Jinko Graham (2018). \strong{SimRVSequences FINISH REFERENCE}.

cite SliM (Haller and Messer).
}
\keyword{datasets}
