% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessStudy_Functions.R
\name{preProcessSNVs_studyDat}
\alias{preProcessSNVs_studyDat}
\title{Process the SNVs in the famStudy object}
\usage{
preProcessSNVs_studyDat(famStudy_obj)
}
\arguments{
\item{famStudy_obj}{An object of class \code{famStudy}.}
}
\value{
An object of class \code{famStudy}. The SNVs in the returned \code{famStudy} object have been filtered based on exclusion criteria.
}
\description{
This willl become an internal function. This is a processing step whereby we filter SNVs for specific exclusion criteria.
}
\details{
In this pre-processing step we filter any variants that are introduced by multiple founders in the same pedigree or that are introduced by the same founder mulitple times as this violates our model assumptions.  Additionally, we remove any SNVs that are not observed in at least one disease-affected relative in the study.
}
\keyword{internal}
