% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPedigreeFunctions.R
\name{processPed}
\alias{processPed}
\title{Process Pedigree}
\usage{
processPed(ped, carriers, subtypes)
}
\arguments{
\item{ped}{An object of class \code{ped}.}

\item{carriers}{subjects in pedigree share the variant}

\item{subtypes}{A list of two subtypes that identify the subtype ranking.  The label for the more the genetically-complelling subtype is listed first.}
}
\value{
A list of information for the pedigree
}
\description{
Process Pedigree
}
\keyword{internal}
