% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Macro_SimFunctions.R
\name{quick_compute_distributions}
\alias{quick_compute_distributions}
\title{Compute all distributions - used for power simulation - avoids haveing to re-run this function multiple times}
\usage{
quick_compute_distributions(peds, subtypes, carrier_prob = 0.002,
  tau_increment = 0.05, file_path, ped_indicies, useK = FALSE)
}
\arguments{
\item{peds}{A ped object containing two or more pedigrees.}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}

\item{carrier_prob}{The cumulative carrier probabilty of all crvs as a group.}

\item{tau_increment}{Numeric. The width of the grid for the taus}

\item{assumption}{The assumption to use for computing the LRs (can be "global" or "semi-global")}
}
\value{
A data frame of global configurations anf weights for the set of provided families.
}
\description{
used for power simulation - avoids haveing to re-run this function multiple times
}
\examples{
library(RVMethods)
data(study_pedigrees)

library(SimRVPedigree)
plot(study_pedigrees[study_pedigrees$FamID == 304, ])
plot(study_pedigrees[study_pedigrees$FamID == 58, ])

peds = study_pedigrees[study_pedigrees$FamID \%in\%c(58, 304), ]
subtypes = c("HL", "NHL")


myCombos = test_allcombos_includeZero(ped = study_pedigrees[study_pedigrees$FamID == 58, ],
                                      subtypes = c("HL", "NHL"),
                                      tau_grid = make_tauGrid(),
                                      carrier_prob = 0.002)

my_GDist = compute_globalDist(peds = study_pedigrees[study_pedigrees$FamID \%in\%c(58, 304), ],
                                subtypes = c("HL", "NHL"))

order(my_GDist$LR) == order(my_GDist$w_LR)

}
\keyword{internal}
