% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyAnalysisFunctions.R
\name{remove_invalidConfigs}
\alias{remove_invalidConfigs}
\title{Remove invalid configurations from the items returned by test_allcombos}
\usage{
remove_invalidConfigs(test_allCombos_result)
}
\arguments{
\item{test_allCombos_result}{The items returned by \code{test_allcombos}, i.e. a list containing \code{like_mat}, \code{configs}, and \code{binIDs}.}
}
\value{
A list containing \code{like_mat}, \code{configs}, and \code{binID}, with the same format as \code{test_allcombos}, but with invalid configurations removed.
}
\description{
Remove invalid configurations from the items returned by test_allcombos
}
\keyword{internal}
