% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_by_obsSNV_Functions.R
\name{SNV_to_config}
\alias{SNV_to_config}
\title{Convert vector of SNV statuses to TRUE/FALSE configuration}
\usage{
SNV_to_config(SNVvec)
}
\arguments{
\item{SNVvec}{a column of the haplotype data matrix}
}
\value{
a vector of T/F statuses
}
\description{
Convert vector of SNV statuses to TRUE/FALSE configuration
}
\examples{
library(SimRVSequences)
data(study_peds)

# construct an object of class SNVdata
ex_data <- SNVdata(Haplotypes = EXhaps,
                   Mutations = EXmuts)

set.seed(11956)
# simulate SNV data with sim_RVstudy
study_seq <- sim_RVstudy(ped_files = study_peds,
                         SNV_data = ex_data)

# determine the class of study seq
class(study_seq)


study_seq$ped_haplos[1:10, 3]

SNV_to_config(study_seq$ped_haplos[1:10, 3])


}
\keyword{internal}
