% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgain_setup_functions.R
\name{TauProbTable}
\alias{TauProbTable}
\title{Table of inheritance probabilities that depend on tau}
\usage{
TauProbTable(tau)
}
\arguments{
\item{tau}{The probability that an offspring inherits the RV from a parent who is a carrier.}
}
\value{
An array of dimension c(3, 3, 3).
}
\description{
Table of inheritance probabilities that depend on tau
}
\keyword{internal}
