% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamAnalysisFunctions.R
\name{test_allcombos}
\alias{test_allcombos}
\title{Compute sharing probabilities for every configuration of disease-affected relatives and initialization of tau}
\usage{
test_allcombos(ped, subtypes, tau_grid, subtype_weights = NULL)
}
\arguments{
\item{ped}{An object of class \code{ped}. A pedigree generated by \code{sim_ped} or \code{sim_RVped}, or an object created by the function \code{new.ped}, for details please see documantation and vignette for package \code{SimRVPedigree}.}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}

\item{tau_grid}{A grid of tau values produced by \code{expand.grid}.  \code{test_allcombos} will compute each sharing probability at every specified value of tau in \code{tau_grid}.}

\item{subtype_weights}{A vector of length 2.  When assuming an informative prior for the founders; the weights the individuals with subtypes A and B, respectively.  By default, \code{subtype_weights = NULL} so that no founder weights are applied (i.e. flat founder prior).}
}
\value{
A list containing the following:

\item{\code{like_mat} }{A matrix of values that represent the likelihood given a configuration (column) and relized value of tau (row).}

\item{\code{configs} }{A matrix of sharing configurations.  The columns are named for the IDs of the disease-affected relatives in the pedigree. When \code{TRUE} the individual is a carrier of the cRV, when \code{FALSE} the individual is not a carrier.}
}
\description{
Compute sharing probabilities for every configuration of disease-affected relatives and initialization of tau
}
\keyword{internal}
