% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Macro_SimFunctions.R
\name{test_allcombos_noCarProb}
\alias{test_allcombos_noCarProb}
\title{Compute sharing probabilities for all configuration (includes the zero configuration) and initialization of tau}
\usage{
test_allcombos_noCarProb(ped, subtypes, tau_grid, subtype_weights = NULL)
}
\arguments{
\item{ped}{An object of class \code{ped}. A pedigree generated by \code{sim_ped} or \code{sim_RVped}, or an object created by the function \code{new.ped}, for details please see documantation and vignette for package \code{SimRVPedigree}.}

\item{subtypes}{A list of length 2. Contains character labels for the two subtypes that occur in the pedigrees contained in the \code{famStudy_obj}. The label for the more the genetically-complelling subtype must be listed first.}

\item{tau_grid}{A grid of tau values produced by \code{expand.grid}.  \code{test_allcombos} will compute each sharing probability at every specified value of tau in \code{tau_grid}.}

\item{subtype_weights}{A vector of length 2.  When assuming an informative prior for the founders; the weights the individuals with subtypes A and B, respectively.  By default, \code{subtype_weights = NULL} so that no founder weights are applied (i.e. flat founder prior).}

\item{carrier_prob}{The cumulative carrier probabilty of all crvs as a group.}
}
\value{
A list containing the following:

\item{\code{like_mat} }{A matrix of values that represent the likelihood given a configuration (column) and relized value of tau (row).}

\item{\code{configs} }{A matrix of sharing configurations.  The columns are named for the IDs of the disease-affected relatives in the pedigree. When \code{TRUE} the individual is a carrier of the cRV, when \code{FALSE} the individual is not a carrier.}
}
\description{
Compute sharing probabilities for every global configuration (includes the zero configuration) and initialization of tau.  The probability of the zero configuration will be used to condition on the event that the RV was observed in at least one study member.  This function requires an assummed value for the carrier probability of the RV
}
\examples{
library(RVMethods)
data(study_pedigrees)

library(SimRVPedigree)
plot(study_pedigrees[study_pedigrees$FamID == 58, ])
myCombos = test_allcombos_includeZero(ped = study_pedigrees[study_pedigrees$FamID == 58, ],
                                      subtypes = c("HL", "NHL"),
                                      tau_grid = make_tauGrid())

myCombos

rowSums(myCombos$like_mat)

}
\keyword{internal}
